Cleanup doc/conf.py.
Emit the latex symbol tables without breaks.
